Feat: 회원탈퇴 시 주문 활성화 되어있는 경우 예외처리

+++ src/main/java/market/demo/service/MemberService.java
@@ -5,6 +5,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import market.demo.domain.member.Member;
 import market.demo.domain.member.jwt.Authority;
+import market.demo.domain.status.OrderStatus;
 import market.demo.dto.MemberDeletionRequest;
 import market.demo.dto.changememberinfo.MemberInfoDto;
 import market.demo.dto.changememberinfo.ModifyMemberInfoDto;
@@ -62,6 +63,15 @@ public class MemberService {
     public void deleteMember(MemberDeletionRequest deletionRequest) {
         Member member = memberRepository.findById(deletionRequest.getMemberId())
                 .orElseThrow(() -> new MemberNotFoundException("멤버를 찾을 수 없습니다. ID: " + deletionRequest.getMemberId()));
+
+        // 주문 상태 확인
+        boolean hasActiveOrders = member.getOrders().stream()
+                .anyMatch(order -> order.getOrderStatus() == OrderStatus.PENDING || order.getOrderStatus() == OrderStatus.PROCESSING);
+
+        if (hasActiveOrders) {
+            throw new IllegalStateException("활성 주문이 존재하여 회원 탈퇴가 불가능합니다.");
+        }
+
         memberRepository.delete(member);
     }
